Add and delete the bone observers properly
Before the bones the armature was listening to were the bones it created.
However, the armature should really be listening to the bone added to it.
(No matter where they come from).

Also, adding the RemoveBoneObservers method. Before that, all the observers
where removed which is not ideal is the bone is used elsewhere. Now only
the armature observersof the bone are removed.

Finally, making the UpdateBoneWithArmatureOption available in the API.
This will help to put the bone on par with the armature options.

Issue #13641