Rest/Pose Mode
Create the appropriate function in the Qt widget to connect the GUI to
the armature node.
In the armature node, a member Widget state was created to hold the
rest/pose state. This state is not hold by the armature node widget
ArmatureProperties because the ArmatureProperties does not have any bones,
making it impossible to hold the Rest/Pose property.

Finally, correcting the displayble manager so it update the widget FROM the
node correctly.

See Issue #13675