Fix Rest/Pose Transition
Before the pose mode was too independent of the rest mode and did not
apply the change of the rest mode configuration properly.

In order to do so, we created the RestToPoseRotation that is computed in
pose mode. It represents the rotation from the local tail position in rest
mode to the local tail position in pose mode. It is basically the
rotation that an user applies to a given bone in pose mode
(with no influence of the parent).

Given this, we can recompute the local pose points and the ParentToBone
rotation (and thus WorldToBone).

See Issue #13681