Proper deletion of the bones
When deleting a bone, we want to delete all its children. For this to
be done correctly, we delete the leaf-bones first and work our way up
to the selected bone for deletion (using recursivity).

This is done by implementing our own "delete" action of the tree view and
disabling the default delete action.

See Issue#13627