Add logic for LinkedWithParent
We decided that it would be easier for the bone node to hold the property
LinkedWithParent since this property needs to be associated to each bone.

The GUI change are rather simple. Just preveting to update the head
coordinates if the bone is linked to its parent.

The only tricky part is to make sure to update the property whenever
 either the armature widget or the bone node has been modified.
This is done in the displayable manager.

See Issue#13627