Improve Armature robustness
Improve the constistency of when the bones are added to the internal
containers (Bones and TopLevelBones).
Breaking the loop when reparenting; this will prevent potential crash and
improve effieciency.

Finally, getting rid of the debugId. Now that bones have names it is not
necessary to have an ID for debugging.

See Issue#13616