Display properties to bone display node
Transfer the display properties to the bone display node.
This implies to udpate the bone display node everytime the bone node is
changed and everytime the widget is changed.

We also now listen to DisplayModifiedEvent() events to be able to update
the widget from the display node.

We must be careful since the display node is created after everything
is setup, we need to update it to the current display.
To this matter we know that a DisplayModifiedEvent() with null calldata is
the first signal sent after the creation of the display node. Upon this
specific event, we initialize the display node.

See Issue#13629