Update GUI when widget is selected
We need to update the gui when the widget is selected so the current bone
in the tree is the selected bone in the API.
To this reason, we use a signal on the armature beacuse the gui listens
to the current armature.
(This prevent us from having the gui listening to all the bones in its
tree).
The event is fired with the modified bone node's ID. This ID is then used
in the GUI to select the current bone.

See Issue #13629