Stop showing parenthood for top-level bones
For the bones without parent, we want to stop showing the parenthood
because it is only a line to the origin.
Fot his, we add a member in vtkMRMLBoneNode to decide if the bone has a
bone parent.
This also helps in the GUI for the linked parent check box and the head
coordinate widget.

See Issue #13698