Refactor initialization of the display bone node
Since the display node is the last to be added to the scene, it must be
properly initialized with the widget values before fullfilling its destiny.
For that we relied on the event DisplayModifiedEvent that is suppose to
fire an event with empty calldata upon creation of the display node.
This led to 2 issues:
- The first DisplayModifiedEvent fired isn't with empty data. This is a
 bug and steered us toward the previous ugly hack.
- It does not work !
Indeed with multiple bones, all the display node can be changed and
and pass through that function. That means that only the first of them
would not update the widget from its values.

Instead, we simply create a ShouldBeInitialized flag for each display
node that can be checked and changed once the node has been initialized.

See Issue #13693