Right Number of click for adding bones
Fist, adding a selected armature node, so we constantly know the currently
selected node in the armature. This assumes that only one armature can
be selected per scene (and consequently by displayable manager).
Secondly, when adding a bone, we check if there is a currently selected
armature and if there is a selected bone in it. If so, we add it on the
first click. Else, the bone is added with two clicks.

See Issue #13724