Enforce representation for armature's node
A bone from an armature now always have a representation. This implies that
the BonesRepresentation can be declared and used for storing the
AlwaysOnTop property.

Also, make sure there are no memory leaks and actually register the
given BonesRepresentation (like said in the comments, but not in the code).

See Issue #13733