Fix ReparentBone()
First, only actually reparent the given bones if necessary:
 - The bone belongs to the armature
 - The new parent belongs to the armature or is null (reparent to root)
 - The bone's parent is different than the new parent

Then the bone was not moved with all its children which is what the user
would expect. It is fixed by adding the children correctly to the new node
and deleting the references from the children

Finally, the method RemoveNodeFromHierarchy does not delete the node.
This still needs to be done after. Note that there is no check here of the
find result since we already know that the old node exists in the armature.

See Issue #13616