Fix MergeBones()
The head parent would not have its children properly update, which would
 lead to a crash when the later was updated.

The new version creates a bone similar to the head bone. Then it reparent
the tail bone to the new bone, leaving the head bone childless.
So when the tail bone is deleted, all its children are reparented
automatically to the new bone. The head bone can be deleted with no regrets
since it has no children.

Finally, the new bone tail is place to the tail bone original tail
position.

See Issue #13616