Display handles last
In X-Ray mode the depth test is disabled; the call order matters
as what is called last will be displayed on top of all the rest.
In order to have a better interaction, we want to display the handles
without artifact from the cones or the cylinder representation.
This is why we ask the cylinder/cones actor to display before the handles.

See Issue #13755