Tweak interaction
Prevent the bone to update itself in PlaceHead mode. It makes no sense
to have rest transform without a tail.
Also, only actie the handles (when the user is hovering) if it's not
the PlaceHead mode in which case nothing at all should be displayed.

Finally clean up the interaction test.

See Issue #13720