Prevent bone parent selection on head
Previously, with linked bones, the active bone would jumpt to the bone
parent when clicking on the bone's head. This wasn't very natural.
To prevent this, the application check when clicking on a bone's tail
that none of its direct children is selected. If one is, then the call
to the event that would transfer the selection is skipped.

See Issue #13710