Clean Up bone node - Display node first
This starts by declaring the bone display node before the node itself.
This is necessary because otherwise the bone node does not have a proper
reference on its display node when loading a scene.

This implies that we can get rid of the initialization-hackish code in
the bone display node and the displayable manager.

See Issue #13668