Envelope radius ratio for each bone.
Instead of having a envelope radius, we have an envelope ratio that defines
a the radius proportionaly is to the bone's length.
Consequently deleting the ratio in the armature dialog so its is not
confused by the user. The envelope visibility has been moved in the bone
section as well. Even though it is an armature property, it would probably
be confusing for the user to have the envelope properties in two different
places.

See Issue #13680