Improve interface
Delete the evaluate weight CLI It is not of any use in the workflow.

Improving the Next/Previous look and feel. They actually have the title
names of the next/previous page. The title is recomputed at every page
based on the toolTip of this page. It is display in a label for now above
the CLIs.

Finally tweaking the layouts margins for a better look.
See Issue #13882