Restore apply button for color table
The color table is only applied when the button is clicked. This will
hopefully prevent misleading the user into believing that the look-up table
is already applied to the node.
Throughout the code, it is implied that if we have a volume node, it has
a display node associated.

See Issue #13882