Use Grayscale Model Maker for skin surface
The best way to have a skin surface is to use the grayscale model maker
and extract the skin from the difference between the body and the outside
(air). This prevent from having any hole in the surface from missing or
"thin" labels, especially when working whit down-sampled labelmaps.

See Issue #13882