Fix Crash
The issue came from the fact that while the node and the bone widget were
correctly updated, the bone representation in the armature widget was not.
This caused the de-synchronization of the node and the widget.
Since the opacity and the color are not (and should not) actively
propagated when changing the armature widget's bone representation, it
needs to be done manually.
The new function UpdateBoneRepresentation() was created for this purpose.

See Issue #13908