Add independent opacity for the envelope
The opacity of the envelope is now set indepently from the opacity of
the bone, through the armature node.
In an effort to make things easier, the opacity of the envelopes is always
lesser or equal to the opacity of the bones (at least when interacting
with the GUI).
The implementation followed the specifity of the bone's opacity and color.

Finally, the widget representation needed to be tweak to be able to handle
geometry that are both transculent and opaque.

See Issue #13680