Speed up armature weight
Since computing the heat diffusion on dense volumes takes a lot of time,
we add the option to use a different spacing on the volumes used for the
computation. The input volumes are downsampled, the heat diffusion is
computed and then the output weight are upsampled.
Downsampling is done using nearest neighbors interpolation and upsampling
using linear interpolation.

See Issue #13944