Delete bones option on armature nodes too
So far, the "Delete Bones" option the the bone tree view only worked on
bone node. However it was also available for armature nodes although it
did nothing.
This changes the behavior of the last case, where it will delete all the
bones of the armature without deleting the armature itself.
The behavior when clicking on a bone is preserved (it will still delete
the current bone and all its children).

See Issue #13923