Stop place mode upon bone/armature deletion
When in persitent mode, the user can still delete armature or/and bones.
This creates a weird state where the user still is in adding mode
although he cannot. To prevent this, evrytime the user deletes a bone or
an armature (or change to another armature), the widget switches back to
normal interaction mode.

See Issue #13923