Output Compute Armature Weight debug images to weight folder
Instead of dumping the debug images in "./" which can cause problems when
executing an installed version of Bender, the images are outputed to the
weight directory, with the prefix corresponding to the given weight.
Example, for weight #3, the weight image is named "weight_03" and the debug
images will be "weight_03_DebugImageName".