Fix restore armature
The previous code used the SetActiveBone and SetActiveArmature which are
not quite the same as setting the active annotation. There can be an
active armature even though the current active annotation is a bone.
This is were the second top-level bone would add itself to the previous
bone rather than the armature; the bone was still the active annotation
node.

See Issue #13974