Prevent heat diffusion in loosely-related bones areas
In the areas that are connected when they should not, the heat will
diffuse thus causing problem when moving the bones.
An example can be if the arm is touching the thigh, the weight for the
thigh should still propagate in the hips and ribs but should not propagate
in the hand.

To this end the heat diffusion is allowed to propagate only in regions
that belong to bones close enough in the family tree. This is controled with
the parenhood distance.
The distance of a bone to one of its parent is retrieved using Djikstra on
the parenthood array of the input armature. If the armature as no parenthood
array, then any bone is related. Otherwise, a bone has a distance of 1 with
its parent and immediate children.

Finaly restrict the re-sampled weight domain by masking it with the body
partition labelmap. This will prevent some of the leaking of the weight
computation outside of the body.

Also harmonize debug images casing.

See Issue #14025