Add Skin Model Maker scripted module
This module create a skin model contained within the given volume.
It really is a wrapper around the ChangeLabel, GrayscaleModelMaker and
ModelQuadricClusteringDecimator.
The principle is to set the volume's skin label to background and then run
the model maker, with the option to decimate it after using the decimator.

See Issue #14026