Add names and pose mode loading capabilities
Previously, the importer could not load poses or the bones names.
The bone naming is pretty straightforward, a "Names" array was added to
the armature polydata, allowing to retreive these names.

For the pose mode it is a little bit trickier. By default, the bone pose
mode is initialized to the bone rest mode the first time the state goes to
pose mode (or when using ResetPoseMode). This makes it easier for the user
because he doesn't have to initialize the pose by hand for each bone.
However in our case, we actualy want to do it. For this, we need to set
the Rest To Pose Rotation. The array is thus added to the armature
polydata.

See Issue #14115