Add preset armature to the armature module
From now on, default armature can be added in the scene. The default
armature are read into the scene using the armature module logic. To add an
new armature resource, one can simply add the corresponding *.vtk file in
the Resources/Armatures directory and modify the CMakeLists accordingly.

See Issue #14024