Integrate Skin Model Maker in workflow
To facilitate the use of the skin model maker, the CreateSkinModel function
is taken out of the SkinModelMakerWidget. The previous skin maker
(a simple grayscale model maker) is replaced by a call to the skin model
maker logic. This defines a vtkMRMLCommandLineModuleNode and updates it
when running in order to expose more or less the same API than the regular
CLIs.

See Issue #14026