Remove harcoded bone and background values from ComputeArmatureWeight
Adding background label value and bone label value as parameters given by
the user.
Also changing the image type used in the ArmatureWeightWriter to the type
it should be (CharImageType). The LabelImageType previously used is the
type of the rest labelmap, but the type of the partition images is
CharImageType.

See Issue #13916