Add bender sample data
This modules internally relies on the sample data module, most notably for
the logic. However, in order to keep the sample data out of bender, the
module is added  to the list of modules to ignore automatically. Thus the
module sample data is never loaded although we can use the code.

The bender sample data is different from the sample data module as it uses
a tree veiw to display the data available. The data in the tree view is
internally mapped to a static dictionnaries that tells the module where to
download the data.