Add BVH file reader
Add a BVH reader to reader motion capture files. To see more reference on
the BVH format itself, see link at the bottom. The BVH reader reads a bvh
file and outputs a polydata file and an armature file. The polydata is in
fact the polydata obtained from the armature.

This lead to a few change in the vtkArmatureWidget and vtkBoneWidget.
In vtkBoneWidget, the RotateTail function was changed to properly update
the bone's transform. Also changing the method name to make sure that
the name is more easily understandable.
In vtkArmatureWidget, add a method to get whether a bone has any child or
not. This prevent from querying all the children just to know if there are
any. Also renaming the IsBoneParent method to IsBoneParentOf for a better
API.
The armature and bone test were updated accordingly.

ref BVH: http://research.cs.wisc.edu/graphics/Courses/cs-838-1999/Jeff/BVH.html