Use a storage node to read in *.vtk and *.arm files
This is the more proper way to import data in slicer. This offloads the
import code from the armature logic.

Warning, this code uses a hack to work.

Background:
When adding an annotation in the scene, the annotation logic responds to
the vtkMRMLScene::NodeAddedEvent and add a hierarchy node to the new
annotation. The hierarchy node's parent is determined by the variables
ActiveAnnotationHierarchyID from the annotation logic.
Setting the right hierarchy node parent is very important as the armature
displayable manager needs them to determine how to build the widget. If the
node hierarchy isn't right, the widget hierarchy will be wrong as well.

What was done before:
When importing an armature, to make sure new bones were added correctly,
the logic would set the ActiveHierarchyNodeID to the appropriate hierarchy
node.

What was the problem:
Two problems can be found. First, using the logic to import file is
definitely not the slicer way of doing things. Secondly, if the user
happened to click at the right time on the wrong bone, the
ActiveAnnotationHierarchyID could change and the newer bone could be added
to the wrong parent. Also, there is no way to create the proper hierarchy
node before the bone is added to the scene since hierarchy nodes need to
refer to a displayable node by ID. (Since the node isn't in the scene yet,
it has no ID). Changing the hierarchy node parent after it was added isn't
really an option as it would imply that we could support re-parenting
capabilities which would require a lot, lot of refactoring/work.

What is the solution proposed:
The solution proposed is a bit hackish. In the storage node, we create a
callback that is used only when importing data. This callback responds to
the vtkMRMLScene::NodeAddedEvent with an higher priority than the
annotation logic and does the job of creating the hierarchy node properly.
When the annotation logic is called, the hierarchy node is already created
and it doesn't need to do any further treatment.