Add support to read animations from BVH files
So far it is very simple, a slider widget is added in the GUI to be able to
change the pose of the bvh armature. The widget is only enabled if the
armature loaded was a bvh.

To make this happen, a few change were made. In the BVH reader API, some
methods were added to allow external application to retrieve either the
ParentToBone rotation at a given frame for a given bone or to apply
directly the frame to a given armature widget.

In slicer, the armature storage node is now referenced by the armature's
model. Armatures can't reference to the storage node themselves since the
class doesn't inherits from vtkMRMLStorableNode. So instead we use the
model to do it. Method were added in the armature to facilitate the access
to the storage node.