Add animation import
Now that the BVH files can be opened in slicer, it would be interresting
to be able to import the animation from one file to an armature in Bender.
This commit aims at creating the infrastucture necessary for this.

To be able to choose the frame we are importing in Bender, an IO option
widget was added. This widget reads the selected file and displays a
preview armature in a popup widget.

The first fix necessary is to store the frame information in the armature
node. Before it was only driven by the value of the GUI which doesn't work
when we can choose what frame is used before touching the armature's
module GUI.

An import button was added in the armature module. It basically only opens
a dialog with the correct io properties so qSlicerArmaturesIO can now that
its dealing with an import rather than a loading.
Right now the import is over simple, it is only based on a name
correspondence between the bone of the armature. More techniques will be
added later.