Fix animation import on non-bvh armature
The animation import only worked on bvh compatible armatures so far.
It is now fixed to work with nonbvh armature although it is still using
the name to find the correspondence between bones which is less than ideal.

Also reseting the camera around the armature bounds in the frame selector
so the whole armature is captured every frame.