Massive refacoring of the workflow to prepare 2.0
Warning: This commits breaks everythings after the steps number 2.

The workflow is refactored to prepare the new bender workflow. The new
worflow has different steps that will depend on different CLIs. So far,
the first step hasn't changed.
The second step does change as the skin and bone model maker are removed.
They will be replaced by the meshing step from which the bone and skin
model will be extracted.