Adapt Change Label for tetrahedral meshing
The tetrahedral meshing needs 4 different labels: background, skin, bones
and muscle (all of which isn't background, skin or bone).
This change introduces a QTable widget that summarize what labels will
be change into what. This is less messy than having a line edit and a
qMRMLColorNodeComboBox for each type of label.
The code was thus refactored to account for these changes. Hopefully it
will be easier to add another type of label in the future if needed.