Keep the input label value for the mesh material
The output tetrahedral mesh obtained has different material property that
depend on the label maps given to it. As of before, these labels where
given by the mesher, thus making it hard to identify the structural
correspondences in the mesh and the input labelmap.
This is done by going through the relabel map and the input image and
mapping the correspondences between the label.
Now, the material array of the output is controled by the input.