Touchups for Create mesh step
Add the input combo box as a labelmap only combobox.
Prevent the MergeLabel module and the CreateMesh module to select their
output upon creation when they doso. This the next module that would react
to the node changed event to create an output with a name based on the
combo box's naming scheme and not the input.