Display only the meaningful statistics
When running without the grayscale node, some statistics aren't meaningful
at all and crowd the interface. Those statistics are now displayed only
when the grayscale volume is present.
Also modified the export to follow make sure that what the user sees is
what he gets.
Finally, the possible statistics have been factorized in the logic to
simplify the maintenance of it. The chart options are automatically filled
with the relevant option after the statistics are computed

SQUASH WITH grayscale option