Fix output mesh dimensions
Previously, the output mesh did not have the same dimensions than the input
image as cleaver does not take into account spacing or origin.
To make sure that the output mesh matches the input image, each of the
mesh points are transformed according to the input image properties. The
transform polydata filter is used after the clean up step to prevent
the processing of unecessary points (air and padded volumes points).

Also, the padding option was removed since for robustness it's better to
always run with padding.