Refactor percentage statistics
Now the percentage statistics are enabled by a checkbox. When enabled, the
user can choose the what label is going to be used for these statistics.

Also adding the Percentage of x volume and Percentage of total volume
exceot x volume where x is the label chosen by the user.