Prevent point merging in clean filter
When merging points in the vtkPolyDataFilter, the output polydata can be
modified and remove tetrahedrons for lines or triangles. This causes issues
when running the FEM Sofa code so we need to make sure that all the cells
are tetrahedrons.
Preventing the point merging still remove all the outside points (in the
air and the padded volume) but keep all the cells tetrahedral.