Add MaterialPropertyReader step to the workflow
Since the material property step does roughly the same thing as evaluate weight
(i.e. add arrays to the mesh) it make sense to put it with the weight step. The title of this step might change tough.
This commit also removes the automatic connections between the pose
surface cli in the next step and the evaluate weight cli of this step. It
doesn't make sense to keep them anymore as the mesh will need both the
material properties AND the weight evaluated to be able to run.