Improve EvalWeight support for outside vertices
Sometimes vertices lay outside of the image physical space they were
generated from. For example, this can happen when there is no "air"
(emptyness) between the data and the image border.
In those case, the no weight is associated to the outside vertices. This
proves problematic however when using SOFA as ALL the vertices need to have
weight associated to it (or SOFA won't work correctly).

To make sure that all these border-cases vertices still get a weight
associated to them, a more robust PhysicalPointToContinuousIndex is used
in the EvalWeight CLI. In most cases it works just like ITK's
TransformPhysicalPointToContinuousIndex. If the voxel computed is outside,
then the voxel used is the closest neighboring voxel.

This also adds a check to make sure that all the vertices of the input
data will be evaluated (and fails if not).