Use Skin Model Maker for the Extract Skin parameter
The SimulatePose CLI needs the skin for collision detection. The skin
from the volume material extractor is incompatible with the SimulatePose
CLI for two reasons. First, it is most probably going both have holes
(since the input data does). Secondly, the resolution for collision
detection is probably going to be too high (thus drastically increasing
computation times).
We consequently revert to the SkinModelMaker that was previously used in
the SimpleWorkflow.