Repair broken cells
The cleaver tet mesh generation sometimes produces points with NaN values.
These points are no good when they are evaluated with the weights or when
used in SOFA.
In order to patch the mesh (while waiting for Cleaver to be fixed), we
make sure that every cell has valid points. To prevent overhead, the lookup
is done when the point and cells are transfered from  cleaver to vtk.
The patching relies on the vtkCenterOfMass.