Fix UNIX animation import
The crash came from the keyword motion that had an extra blank character,
thus being unrecognized as the right keyword. This would cause the reader
to become invalid, deleting the armature. This was a problem when copying
the armature's polydata and causing the crash.
First, improve the error message in ParseMotions. Also check for the
ApplyFrameToArmature return value so the reader can fail more gracefully.

Finally, to be sure to get rid of all the blank when looking for a keyword,
the STD's function isspace is used instead of the homemade IsBlank.