Uniformize input and output directories throughout modules
By using settings, we can uniformize the use of "Bender" directories
in the application. This is a major help in the workflow as will stop the
user to have to change directory every time something needs to be saved.

First, the settings are set in the BenderWelcome if they don't already
exist. This helps to have consistent behavior even if it's the first time
the application is started.
In BenderSampleData, the InputDirectory setting is used to know where the
data can be downloaded.
Finally in the workflow (FEM Workflow), all the loading/saving was
customized to use the InputDirectory/OutputDirectory setting.