Allow interaction with the octoherdorn or cylinder in pose mode
So far in pose mode the only way to move the bone was to either click on
the line or the tail. The user should also be able to interact with the
octohedron or the cylinder in pose mode.
Code-wise, the pose mode check was simply remove to allow interaction with
the octohedron or cylinder.