Fix pose mode bug
In pose mode it can happen that the bone currently manipulated and all its
children will disappear. This was caused by the lines used for the rotation
computation being the same, leading to a bad angle (NaN in this case).
The event where the two lines are the same are now simply disregarded.
An assert was added to help for future debugging.