Fix VotingResample border issue
Because the radius was constant, the output image wouldn't necesseraly have
the right label when resampling with a high spacing. It also assumed that
the image was uniformely spaced in 3D which might not be the case.
This fix bases the radius on the output image spacing thus taking into
account all the necessary voxel for voting.