Simplify workflow's resample spacing
Instead of a coordinates widget box, we can use instead a simple spinbox
that show the scale with which the image spacing can be resampled. This
echoes to what is done for the compute weight step.
Also, since the spacing adjustement is done is the CLI now, we can simply
delete what was done here.
Finaly, fix the go to voting resample.