Fix armature construction
The problem was the same than for the vtkMRMLArmatureStorageNode. The
hierarchy node is added too late and the widget has already figured that
the bone node shouldn't be displayed. This causes issues as the bone added
isn't correctly selected.
The solution is the same than what was done for the
vtkMRMLArmatureStorageNode. It is actually the same code, simply moved to
a central place in the vtkMRMLBoneNode.
This is still a hack and a refactoring of the bone/armature selection
would be preferrable.