Refactor the pose animation widget
Refactor the pose part of the armature widget in its own stand-alone widget
so it can easily be re-used in the workflow.
Right now the API of that widget isn't particurarly developped or up to the
standards of a stand-alone widget but it's enough for the very specific
tasks it has.