Fix animation import for the arm
When doing the initial resizing of bones, it used the local tail to
determine the resizing size. While this works for all the bones, this
doesn't work in the case where both the animation root head isn't at the
origin. In that particular case, the distance would be the regular bone
distance.